Aplicando patron strategy
