--- InvalidSalaryException.java
@@ -1,7 +1,0 @@
-
-public class InvalidSalaryException extends Exception {
-
-    public InvalidSalaryException(String message) {
-        super(message);
-    }
-}
