+++ src/employee/management/InvalidSalaryException.java
@@ -1,0 +1,9 @@
+package employee.management;
+
+
+public class InvalidSalaryException extends Exception {
+
+    public InvalidSalaryException(String message) {
+        super(message);
+    }
+}
